# expression: implement vectorized evaluation for builtinRegexpSig