--- a/expression/builtin_like.go
+++ b/expression/builtin_like.go
@@ -156,9 +156,13 @@ func (b *builtinRegexpSig) evalInt(row chunk.Row) (int64, bool, error) {
 	}
 
 	// TODO: We don't need to compile pattern if it has been compiled or it is static.
-	re, err := regexp.Compile("(?i)" + pat)
+	re, err := b.compilePattern(pat)
 	if err != nil {
 		return 0, true, ErrRegexp.GenWithStackByArgs(err.Error())
 	}
 	return boolToInt64(re.MatchString(expr)), false, nil
+}
+
+func (b *builtinRegexpSig) compilePattern(pat string) (*regexp.Regexp, error) {
+	return regexp.Compile("(?i)" + pat)
 }