--- a/expression/builtin_like.go
+++ b/expression/builtin_like.go
@@ -96,9 +96,9 @@ func (c *regexpFunctionClass) getFunction(ctx sessionctx.Context, args []Express
 	bf.tp.Flen = 1
 	var sig builtinFunc
 	if types.IsBinaryStr(args[0].GetType()) || types.IsBinaryStr(args[1].GetType()) {
-		sig = &builtinRegexpBinarySig{baseBuiltinFunc: bf, rm: nil}
+		sig = &builtinRegexpBinarySig{builtinRegexpSharedSig{baseBuiltinFunc: bf}}
 	} else {
-		sig = &builtinRegexpSig{baseBuiltinFunc: bf, rm: nil}
+		sig = &builtinRegexpSig{builtinRegexpSharedSig{baseBuiltinFunc: bf}}
 	}
 	return sig, nil
 }
@@ -107,31 +107,14 @@ type regexpCompiler interface {
 	compile(pat string) (*regexp.Regexp, error)
 }
 
-type regexpMemoizer struct {
-	re *regexp.Regexp
-}
-
-func newRegexpMemoizer(patterns *chunk.Column, n int, rc regexpCompiler) *regexpMemoizer {
-	// Precondition: patterns is generated from a constant expression
-	rm := &regexpMemoizer{}
-	for i := 0; i < n; i++ {
-		if patterns.IsNull(i) {
-			continue
-		}
-		re, err := rc.compile(patterns.GetString(i))
-		if err == nil {
-			// Because patterns rows are all the same, we memoize the first
-			// successfully compiled regexp
-			rm.re = re
-			break
-		}
-	}
-	return rm
+type builtinRegexpSharedSig struct {
+	baseBuiltinFunc
+	memoizedRegexp *regexp.Regexp
+	memoizeErr     error
 }
 
 type builtinRegexpBinarySig struct {
-	baseBuiltinFunc
-	rm *regexpMemoizer // only used in vectorized evaluation
+	builtinRegexpSharedSig
 }
 
 func (b *builtinRegexpBinarySig) Clone() builtinFunc {
@@ -141,6 +124,7 @@ func (b *builtinRegexpBinarySig) Clone() builtinFunc {
 }
 
 func (b *builtinRegexpBinarySig) evalInt(row chunk.Row) (int64, bool, error) {
+	// TODO: this can also be moved into builtinRegexpSharedSig
 	expr, isNull, err := b.args[0].EvalString(b.ctx, row)
 	if isNull || err != nil {
 		return 0, true, err
@@ -164,8 +148,7 @@ func (b *builtinRegexpBinarySig) compile(pat string) (*regexp.Regexp, error) {
 }
 
 type builtinRegexpSig struct {
-	baseBuiltinFunc
-	rm *regexpMemoizer // only used in vectorized evaluation
+	builtinRegexpSharedSig
 }
 
 func (b *builtinRegexpSig) Clone() builtinFunc {
@@ -177,6 +160,7 @@ func (b *builtinRegexpSig) Clone() builtinFunc {
 // evalInt evals `expr REGEXP pat`, or `expr RLIKE pat`.
 // See https://dev.mysql.com/doc/refman/5.7/en/regexp.html#operator_regexp
 func (b *builtinRegexpSig) evalInt(row chunk.Row) (int64, bool, error) {
+	// TODO: this can also be moved into builtinRegexpSharedSig
 	expr, isNull, err := b.args[0].EvalString(b.ctx, row)
 	if isNull || err != nil {
 		return 0, true, err