--- a/expression/builtin_like_vec.go
+++ b/expression/builtin_like_vec.go
@@ -34,18 +34,6 @@ func (b *builtinRegexpBinarySig) vectorized() bool {
 }
 
 func (b *builtinRegexpBinarySig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) error {
-	return vecEvalRegexp(&b.baseBuiltinFunc, b, input, result)
-}
-
-func (b *builtinRegexpSig) vectorized() bool {
-	return true
-}
-
-func (b *builtinRegexpSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) error {
-	return vecEvalRegexp(&b.baseBuiltinFunc, b, input, result)
-}
-
-func vecEvalRegexp(b *baseBuiltinFunc, rc regexpCompiler, input *chunk.Chunk, result *chunk.Column) error {
 	n := input.NumRows()
 	bufExpr, err := b.bufAllocator.get(types.ETString, n)
 	if err != nil {
@@ -65,31 +53,14 @@ func vecEvalRegexp(b *baseBuiltinFunc, rc regexpCompiler, input *chunk.Chunk, re
 		return err
 	}
 
-	pat2re := make(map[string]*regexp.Regexp)
-	isMemoizable := b.args[1].ConstItem()
-	if isMemoizable {
-		// If args[1] is a constant item, then all the items in bufPat are
-		// the same. So the pattern can be compiled once and memoized.
-		for i := 0; i < n; i++ {
-			if bufPat.IsNull(i) {
-				continue
-			}
-			pat := bufPat.GetString(i)
-			if _, ok := pat2re[pat]; ok {
-				continue
-			}
-			re, err := rc.compile(pat)
-			if err != nil {
-				return err
-			}
-			pat2re[pat] = re
-		}
+	if isRegexpMemoizable(b.args[1]) && b.rm == nil {
+		b.rm = newRegexpMemoizer(bufPat, n, b)
 	}
 	getRegexp := func(pat string) (*regexp.Regexp, error) {
-		if isMemoizable {
-			return pat2re[pat], nil
+		if b.rm != nil && b.rm.re != nil {
+			return b.rm.re, nil
 		}
-		return rc.compile(pat)
+		return b.compile(pat)
 	}
 
 	result.ResizeInt64(n, false)
@@ -106,4 +77,62 @@ func vecEvalRegexp(b *baseBuiltinFunc, rc regexpCompiler, input *chunk.Chunk, re
 		i64s[i] = boolToInt64(re.MatchString(bufExpr.GetString(i)))
 	}
 	return nil
+}
+
+func (b *builtinRegexpSig) vectorized() bool {
+	return true
+}
+
+func (b *builtinRegexpSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) error {
+	n := input.NumRows()
+	bufExpr, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(bufExpr)
+	if err := b.args[0].VecEvalString(b.ctx, input, bufExpr); err != nil {
+		return err
+	}
+
+	bufPat, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(bufPat)
+	if err := b.args[1].VecEvalString(b.ctx, input, bufPat); err != nil {
+		return err
+	}
+
+	if isRegexpMemoizable(b.args[1]) && b.rm == nil {
+		b.rm = newRegexpMemoizer(bufPat, n, b)
+		// Note that the existence of b.rm only means that we have initialized
+		// the regexp memoizer itself. b.rm.re could still be nil at this point,
+		// either because patterns are all NULL, or none of the patterns was
+		// compiled successfully.
+	}
+	getRegexp := func(pat string) (*regexp.Regexp, error) {
+		if b.rm != nil && b.rm.re != nil {
+			return b.rm.re, nil
+		}
+		return b.compile(pat)
+	}
+
+	result.ResizeInt64(n, false)
+	result.MergeNulls(bufExpr, bufPat)
+	i64s := result.Int64s()
+	for i := 0; i < n; i++ {
+		if result.IsNull(i) {
+			continue
+		}
+		re, err := getRegexp(bufPat.GetString(i))
+		if err != nil {
+			return err
+		}
+		i64s[i] = boolToInt64(re.MatchString(bufExpr.GetString(i)))
+	}
+	return nil
+}
+
+func isRegexpMemoizable(argPat Expression) bool {
+	return argPat.ConstItem()
 }