<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Implement vectorized evaluation for `builtinRegexpSig`. Issue: #12103

### What is changed and how it works?

```
BenchmarkVectorizedBuiltinLikeFunc/builtinRegexpSig-VecBuiltinFunc-4         	     500	   3095383 ns/op	 1859769 B/op	   12727 allocs/op
BenchmarkVectorizedBuiltinLikeFunc/builtinRegexpSig-NonVecBuiltinFunc-4      	     500	   3103907 ns/op	 1862208 B/op	   12725 allocs/op
```

Unfortunately, vectorization did not boost the performance in the benchmark test. According to the profiling, `regexp.Compile()` was a major bottleneck in the CPU time.

Given that the regex pattern (`args[1]`) is most likely a string literal in production usage, we optimize for this case -- if `args[1]` is constant, then the pattern is pre-compiled once and memoized. See the manual test below for the details.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Manual test (add detailed scripts or steps below)

I generated a simple table of 100,000 rows with the schema below:

```
id INT, str VARCHAR(32), str2 VARCHAR(32)
```

then ran the following query:

```
SELECT str REGEXP "(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)" FROM test;
```

The vectorized implementation took ~1.2s, while the non-vectorized implementation took ~1.8s. I believe the performance gain is proportional to the complexity of the regex pattern.